requirements.txt
matplotlib==3.0.2
numpy==1.15.4
Pillow==5.3.0